refactor : 오디오앨범 좋아요 코드 정리
컨트롤러와 서비스 분리
